Updated build process, maybe fixed issue 1
Changes to the wscript files to work with newer waf on Python 3.  This
allows me to build the software at least on RHEL 9, right out of the
box.

While I'm at it, I added space around the literal in gui.h on line 171.
This may or may not be the only problem with strict C++11, but it's the
one mentioned in the issue upstream.